Upgrade to d3 v3.5.5
The event handling code has been changed in d3 since the first release of
this plugin. This commit brings the d3.svg.multibrush plugin in sync with
d3 v3.5.5.

There is one issue though: the meta key is not working. I.e. a second brush
will be set (in stead of replacing the former brushes) even when the meta
key is not pressed.